Restructure intro dialogs
