+++ app/src/main/java/org/simonolander/lambda/domain/Exercises.kt
@@ -2,12 +2,27 @@ package org.simonolander.lambda.domain
 
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.withStyle
-import org.simonolander.lambda.content.dialog.booleanDialog
 import org.simonolander.lambda.engine.*
+import org.simonolander.lambda.misc.lambdaCalculus
 import org.simonolander.lambda.ui.theme.codeStyle
 
 val trueExercise = run {
     val functionName = "true"
+
+    val dialog = run {
+        DialogBuilder()
+            .message("Hello!")
+            .message("In this lesson we're going to be exploring booleans in $lambdaCalculus.")
+            .message("A boolean is a type of data that can be one of two values: true or false.")
+            .message("Booleans are used primarily for deciding between two outcomes.")
+            .message("Like everything else in $lambdaCalculus, we represent booleans using functions.")
+            .message("A boolean in $lambdaCalculus is simply a function, that when presented with two arguments, returns one of them.")
+            .message("The boolean true chooses the first argument. False chooses the second.")
+            .message("Please go ahead and write the true function for me.")
+            .message("I'll give you a hint, you've already seen this function in a previous lesson.")
+            .build()
+    }
+
     Exercise(
         name = "True",
         description = buildAnnotatedString {
@@ -44,7 +59,7 @@ val trueExercise = run {
                 output = a,
             )
         },
-        dialog = booleanDialog,
+        dialog = dialog,
     )
 }
